dom/nodes/insertion-effects/Node-appendChild-script-and-stylesheet-link-from-fragment.tentative.html

<!doctype html>
<meta charset=utf-8>
<title>Node.appendChild: inserting script and stylesheet link from a fragment</title>
<script src=/resources/testharness.js></script>
<script src=/resources/testharnessreport.js></script>
<script>
const happened = [];
const link = document.createElement("link");
test(() => {
  link.rel = "stylesheet";
  link.href = "data:text/css,";

  const script = document.createElement("script");
  script.textContent = `
    happened.push(link.sheet === null);
`;

  const df = document.createDocumentFragment();
  df.appendChild(script);
  df.appendChild(link);

  assert_array_equals(happened, []);
  document.body.appendChild(df);
  assert_array_equals(happened, [true]);
});
</script>
